Allow zhack label repair to restore detached devices.
This commit expands on the zhack label repair command in d04b5c9 by
regenerating uberblocks in addition to the existing functionality of
fixing checksums. The changes are heavily inspired by Jeff Bonwick's
labelfix utility, as archived at:

https://gist.github.com/jjwhitney/baaa63144da89726e482

Additionally, it is now capable of properly determining the size of
block devices and other media, as well as handling sizes which are
not divisible by 2^18. This should make it viable for use on physical
devices and partitions, in addition to files.

These changes should make it possible to import zpools that have had
their uberblocks erased, such as in the case of pools rendered
inaccessible by erroneous detach commands.

Signed-off-by: buzzingwires <buzzingwires@outlook.com>